@@ -1,16 +1,16 @@
 {
-  "name": "sokos-okonomiportalen-frontend",
+  "name": "sokos-op-fasade",
   "version": "1.0.0",
-  "description": "sokos-okonomiportalen-frontend",
-  "main": "index.js",
+  "description": "sokos-op-fasade",
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build --base=/okonomiportalen/ --outDir dist",
+    "build": "tsc && vite build",
     "preview": "vite preview",
+    "test": "vitest",
     "prepare": "husky install",
-    "lint": "eslint --fix --ext .js,.jsx .",
-    "lint:staged": "eslint -c .eslintrc.pre-commit.js --fix --ext .js,.jsx ."
+    "lint": "eslint --fix --ext .ts,.tsx,.js,.jsx .",
+    "pretty-quick": "pretty-quick --pattern \"**/*.*(ts|tsx)\""
   },
   "author": "navikt",
   "license": "MIT",
@@ -19,41 +19,52 @@
     "@navikt/ds-css-internal": "^2.1.1",
     "@navikt/ds-react": "^2.1.1",
     "@navikt/ds-react-internal": "^2.1.1",
-    "@navikt/ds-tokens": "2.1.1",
-    "@sentry/react": "7.34.0",
-    "@sentry/tracing": "7.34.0",
+    "@navikt/ds-tokens": "^2.1.1",
+    "@rollup/plugin-typescript": "^11.0.0",
+    "@sentry/react": "^7.34.0",
+    "@sentry/tracing": "^7.34.0",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "amplitude-js": "^8.21.4",
-    "js-cookie": "3.0.1",
-    "react-query": "3.39.3",
-    "react-router": "6.8.0",
-    "react-router-dom": "6.8.0",
-    "rollup-plugin-terser": "7.0.2",
-    "vite": "4.0.4",
-    "vite-plugin-mock": "2.9.6",
-    "zustand": "4.3.2"
+    "react-error-boundary": "^3.1.4",
+    "react-query": "^3.39.3",
+    "react-router": "^6.8.0",
+    "react-router-dom": "^6.8.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "vite": "^4.0.4",
+    "vite-plugin-mock": "^2.9.6",
+    "zustand": "^4.3.2"
   },
   "devDependencies": {
-    "@eik/rollup-plugin": "4.0.40",
-    "@vitejs/plugin-react": "3.0.1",
-    "eslint": "^8.32.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.1",
-    "husky": "8.0.3",
+    "@eik/rollup-plugin": "^4.0.45",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@vitejs/plugin-react": "^3.0.1",
     "autoprefixer": "^10.4.13",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-standard-with-typescript": "^34.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.32.2",
+    "husky": "^8.0.3",
     "lint-staged": "^13.1.0",
-    "prettier": "^2.8.3",
     "postcss": "^8.4.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "^3.2.4"
+    "prettier": "^2.8.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "source-map-loader": "^4.0.1",
+    "tailwindcss": "^3.2.4",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.5"
   },
   "prettier": {
     "printWidth": 120
   },
   "lint-staged": {
-    "*.{js,jsx}": [
-      "eslint -c .eslintrc.pre-commit.js --fix",
+    "*.{js,jsx,ts,tsx}": [
+      "eslint -c .eslintrc.pre-commit.cjs --fix",
       "prettier --cache --write"
     ],
     "*.{css,md,html,json}": [
