@@ -1,10 +1,6 @@
+//Legg inn regler som ikke skal kjøres før commit.
+
 module.exports = {
   extends: [".eslintrc.cjs"],
-  rules: {
-    //TODO: enable rules
-    "react/prop-types": "warn",
-    "no-unused-vars": "warn",
-    "no-undef": "warn",
-    "react/display-name": "warn",
-  },
+  rules: {},
 };
