@@ -1,8 +1,7 @@
-import React from "react";
 import { Loader } from "@navikt/ds-react";
 import ContentLoaderCSS from "./ContentLoader.module.css";
 
-const ContentLoader = () => {
+const ContentLoader = (): JSX.Element => {
   return (
     <div className={ContentLoaderCSS.contentloader}>
       <Loader transparent title="Laster inn..." size="2xlarge" />
